+++ src/test/java/com/configreader/utils/PropertyConfigTest.java
@@ -2,6 +2,7 @@ package com.configreader.utils;
 
 import org.testng.annotations.Test;
 
+import java.util.Map;
 import java.util.Properties;
 
 import static org.hamcrest.CoreMatchers.equalTo;
@@ -12,16 +13,16 @@ public class PropertyConfigTest {
 
     @Test
     public void loadGlobalProps() {
-        Properties properties = new PropertyConfig().getConfig();
+        Map<String, String> properties = new PropertyConfig().getConfig();
         assertThat(3, is(equalTo(properties.size())));
-        assertThat("global", is(equalTo(properties.getProperty("property"))));
+        assertThat("global", is(equalTo(properties.get("property"))));
     }
 
     @Test
     public void loadMergedProps() {
-        Properties properties = new PropertyConfig("specific").getConfig();
+        Map<String, String> properties = new PropertyConfig("specific").getConfig();
         assertThat(5, is(equalTo(properties.size())));
-        assertThat("specific", is(equalTo(properties.getProperty("property"))));
+        assertThat("specific", is(equalTo(properties.get("property"))));
     }
